requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
pymongo==3.12.0
